transferToImageBitmap should throw if there are opened layers
OffscreenCanvas.transferToImageBitmap is incompatible with how the 2D
canvas is rasterized when it contains unclosed layers. Because layers
can have filters that get applied on their final content, they can't
be presented until they are closed. Instead, we normally keep the
layer content alive after a flush, so that it can be presented in
a later frame when the layer is finally closed.

OffscreenCanvas.transferToImageBitmap however is supposed to release
the canvas content, leaving the offscreen canvas empty. We cannot
release the recording if layers are incomplete, and if we kept the
layer content alive for later, we would not be leaving the canvas
empty as the spec requires.